当たり判定と処理完了

+++ Source/Easing.h
@@ -12,8 +12,8 @@ float AttackEasing(float x)
 	const float c2 = c1 * 1.525f;
 
 	return x < 0.5
-		? (pow(2 * x, 2) * ((c2 + 1) * 2 * x - c2)) / 2
-		: (pow(2 * x - 2, 2) * ((c2 + 1) * (x * 2 - 2) + c2) + 2) / 2;
+		? float((pow(2 * x, 2) * ((c2 + 1) * 2 * x - c2)) / 2)
+		: float((pow(2 * x - 2, 2) * ((c2 + 1) * (x * 2 - 2) + c2) + 2) / 2);
 }
 
 float PushEasing(float x)
@@ -25,12 +25,12 @@ float PushEasing(float x)
 	{
 		c1 = 10.0f;
 		c2 = c1 * 1.525f;
-		return (pow(2 * x, 2) * ((c2 + 1) * 2 * x - c2)) / 2;
+		return float((pow(2 * x, 2) * ((c2 + 1) * 2 * x - c2)) / 2);
 	}
 	else
 	{
 		c1 = 1.70158f;
 		c2 = c1 * 1.525f;
-		return (pow(2 * x - 2, 2) * ((c2 + 1) * (x * 2 - 2) + c2) + 2) / 2;
+		return float((pow(2 * x - 2, 2) * ((c2 + 1) * (x * 2 - 2) + c2) + 2) / 2);
 	}
 }
